openf1_data_uploader/pyproject.toml
[project]
name = "data-uploader"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=3.0.2,<4.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
